![image](https://user-images.githubusercontent.com/52546898/231876002-e3a304b4-54a5-473d-b333-8bbf6b12987c.png)

![image](https://user-images.githubusercontent.com/52546898/231876101-ab19a45d-d234-4b44-a1ed-13ffd60da48a.png)

![image](https://user-images.githubusercontent.com/52546898/231876246-e466b808-c26d-4901-adb2-150acc0be2c4.png)

+++ backend/components/auth/deleteuser.js
@@ -16,22 +16,30 @@ const groups = db.collection('groups');
  * Delete the user account
  * @param {string} user_id - The user_id of the user having their password updated
  */
-async function deleteAccount(user_id) {
-  await users.doc(user_id).get().then((doc) => {
-    if (doc.data().groups !== undefined) {
-      doc.data().groups.forEach(async (group_id) => {
-        await group.leaveGroup(user_id, group_id);
+async function deleteAccount(user_id, password) {
+  console.log(password)
+  await users.doc(user_id).get().then(async (doc) => {
+    if (doc.data().password !== password) {
+      throw new Error(401);
+    }
+    else {
+      await users.doc(user_id).get().then((doc) => {
+        if (doc.data().groups !== undefined) {
+          doc.data().groups.forEach(async (group_id) => {
+            await group.leaveGroup(user_id, group_id);
+          });
+        }
+      }).catch((err) => {
+        throw new Error(500);
+      });
+      await users.doc(user_id).delete().catch((err) => {
+        throw new Error(500);
+      });
+      await schedules.doc(user_id).delete().catch((err) => {
+        throw new Error(500);
       });
     }
-  }).catch((err) => {
-    throw new Error(500);
   });
-  await users.doc(user_id).delete().catch((err) => {
-    throw new Error(500);
-  });
-  await schedules.doc(user_id).delete().catch((err) => {
-    throw new Error(500);
-  });
-}
+  }
 
-module.exports = {deleteAccount}
+module.exports = { deleteAccount }
